refactor: apply kebab-case naming policy to React components
