+++ app/components/lp-templates/minimal-template.tsx
