+++ app/components/lp-templates/newspaper-template.tsx
